requirements-hw.txt
pillow>=7.0.0
numpy>=1.16.4,<1.17
requests>=2.20.0
opencv-python>=4.2.0.32
tqdm>=4.23.0
beautifulsoup4>=4.6.0
diffimg==0.2.3
tensorflow>=1.13.1,<1.14
matplotlib>=3.0.2
seaborn>=0.9.0
scikit-image
sklearn
pyyaml
bitstring==3.1.9
jaconv==0.3

#python run.py -l ja -hw -ni 160 -i "../../ligua/TrainHWJapanese/Data/Subject/Address/address.csv" -c 1947
python run.py -l ja -hw -ni 1 -e 'png' -i "texts/random_serial.txt" -c 10000
